Assistant checkpoint: JSONB 쿼리에 타입 캐스팅 추가
Assistant generated file changes:
- src/services/paymentService.js: pending_payments 업데이트 쿼리에 명시적 타입 캐스팅 추가

---

User prompt:

❌ 결제 서비스: 토스 결제 승인 실패: error: could not determine data type of parameter $1
    at /home/runner/workspace/node_modules/pg/lib/client.js:545:17
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
    at async PaymentService.confirmTossPayment (/home/runner/workspace/src/services/paymentService.js:367:9)
    at async confirm (/home/runner/workspace/src/controllers/tossController.js:86:22) {
  length: 124,
  severity: 'ERROR',
  code: '42P18',
  detail: undefined,
  hint: undefined,
  position: undefined,
  internalPosition: undefined,
  internalQuery: undefined,
  where: undefined,
  schema: undefined,
  table: undefined,
  column: undefined,
  dataType: undefined,
  constraint: undefined,
  file: 'postgres.c',
  line: '744',
  routine: 'pg_analyze_and_rewrite_varparams'

  무슨 오류야?

+++ src/services/paymentService.js
@@ -366,7 +366,7 @@ class PaymentService {
         // 실제 생성된 orders.id도 기록
         await updateClient.query(`
           UPDATE pending_payments 
-          SET order_data = order_data || jsonb_build_object('actual_order_id', $1)
+          SET order_data = order_data || jsonb_build_object('actual_order_id', $1::integer)
           WHERE order_id = $2
         `, [result.orderId, orderId]);
         console.log(`✅ 실제 주문 ID ${result.orderId}를 pending_payments에 기록`);
